test: Set minimum on transitive dependency
As part of our testing procedure, we test the SDK with the lowest and
highest allowable versions in our `composer.json` file.

`OpenFeature` has a dependency on `myclabs/php-enum` which has a minor
version of `1.8.0`. In `1.8.1`, a fix was released to address some
return type mismatch from the 8.x update.

Therefore, while OpenFeature can claim to work with `1.8.x` of that
class, we can only really work with `1.8.1+`.